+++ apps/client/src/components/home/Header/Header.tsx
@@ -1,9 +1,9 @@
-import { SpotifyEmbed } from "@modules/SpotifyEmbed";
 import { Typography, useMediaQuery } from "@mui/material";
 import clsx from "clsx";
 import { FC } from "react";
-import { useAppContext } from "src/core/contexts";
+import { useAppContext } from "src/contexts";
 import { theme } from "src/theme";
+import { SpotifyEmbed } from "../SpotifyEmbed";
 
 export const Header: FC = () => {
     const { selectedSong } = useAppContext();
