+++ apps/client/src/components/home/SearchBar/index.ts
