+++ apps/client/src/components/home/SongCard/LyricsCard.tsx
