+++ apps/client/src/components/home/SpotifyEmbed/SpotifyEmbed.tsx
@@ -1,7 +1,7 @@
 import { Slide, useMediaQuery } from "@mui/material";
 import clsx from "clsx";
 import { FC } from "react";
-import { useAppContext } from "src/core/contexts";
+import { useAppContext } from "src/contexts";
 import { theme } from "src/theme";
 
 export const SpotifyEmbed: FC = () => {
