+++ apps/client/src/contexts/appProvider.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { FC, PropsWithChildren, useCallback, useEffect, useState } from "react";
 
 import { fetchAnkiCards } from "../apis/ankiApi";
