--- apps/client/src/pages/_document.tsx
@@ -1,47 +1,0 @@
-import { Head, Html, Main, NextScript } from "next/document";
-
-export default function Document() {
-    return (
-        <Html lang="en">
-            <Head>
-                <link
-                    rel="apple-touch-icon"
-                    sizes="180x180"
-                    href="/apple-touch-icon.png"
-                />
-                <link
-                    rel="icon"
-                    type="image/png"
-                    sizes="32x32"
-                    href="/favicon-32x32.png"
-                />
-                <link
-                    rel="icon"
-                    type="image/png"
-                    sizes="16x16"
-                    href="/favicon-16x16.png"
-                />
-                <link rel="manifest" href="/site.webmanifest" />
-                <link
-                    rel="mask-icon"
-                    href="/safari-pinned-tab.svg"
-                    color="#5bbad5"
-                />
-                <meta name="msapplication-TileColor" content="#da532c" />
-                <meta name="theme-color" content="#ffffff" />
-                <link
-                    href="https://fonts.googleapis.com/css?family=Inter:r,b&display=swap"
-                    rel="stylesheet"
-                />
-                <link
-                    href="https://fonts.googleapis.com/css?family=IBM+Plex+Sans+Thai:r,b&display=swap"
-                    rel="stylesheet"
-                />
-            </Head>
-            <body>
-                <Main />
-                <NextScript />
-            </body>
-        </Html>
-    );
-}
